pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enterprise-deal-analyzer"
version = "0.1.0"
description = "Air-gapped RAG pipeline for financial document analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Senior Architect", email = "architect@example.com" }
]
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    "ragas>=0.1.0",
    "streamlit>=1.38.0",
    "pypdf>=4.3.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "requests>=2.32.0",
    "datasets>=2.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.ruff]
line-length = 88
target-version = "py311"
